Move parser from framework to typed, keep the tests external.
